spotify models refactor
