+++ src/main/java/pl/poznan/put/cs/project/spotifypartyplanner/spotify/model/AuthorizationResponse.java
@@ -1,0 +1,5 @@
+package pl.poznan.put.cs.project.spotifypartyplanner.spotify.model;
+
+public class AuthorizationResponse extends SearchResponse {
+    public String accessToken;
+}
