+++ src/main/java/pl/poznan/put/cs/project/spotifypartyplanner/spotify/model/SearchResponse.java
@@ -1,0 +1,9 @@
+package pl.poznan.put.cs.project.spotifypartyplanner.spotify.model;
+
+public class SearchResponse extends SpotifyResponse {
+    public Tracks tracks;
+    public Tracks getTracks() {
+        return tracks;
+    }
+}
+
