+++ src/main/java/pl/poznan/put/cs/project/spotifypartyplanner/spotify/model/SpotifyResponse.java
@@ -1,0 +1,10 @@
+package pl.poznan.put.cs.project.spotifypartyplanner.spotify.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public abstract class SpotifyResponse {
+}
